Extract Tauri commands into semi-organized modules
