+++ crates/tauri-app/src/commands/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod discover;
+pub(crate) mod manager;
+pub(crate) mod settings;
+pub(crate) mod system;
